+++ src/network/blocking/ServerImpl.h
@@ -23,7 +23,7 @@ public:
     ~ServerImpl();
 
     // See Server.h
-    void Start(uint32_t port, uint16_t workers) override;
+    void Start(uint16_t port, uint16_t workers) override;
 
     // See Server.h
     void Stop() override;
@@ -61,7 +61,7 @@ private:
     // Port to listen for new connections, permits access only from
     // inside of accept_thread
     // Read-only
-    uint32_t listen_port;
+    uint16_t listen_port;
 
     // Mutex used to access connections list
     std::mutex connections_mutex;
