+++ src/network/nonblocking/ServerImpl.h
@@ -22,7 +22,7 @@ public:
     ~ServerImpl();
 
     // See Server.h
-    void Start(uint32_t port, uint16_t workers) override;
+    void Start(uint16_t port, uint16_t workers) override;
 
     // See Server.h
     void Stop() override;
@@ -34,7 +34,7 @@ private:
     // Port to listen for new connections, permits access only from
     // inside of accept_thread
     // Read-only
-    uint32_t listen_port;
+    uint16_t listen_port;
 
     // Thread that is accepting new connections
     std::vector<Worker> workers;
