+++ src/network/uv/ServerImpl.h
@@ -23,7 +23,7 @@ public:
     ~ServerImpl();
 
     // See Server.h
-    void Start(uint32_t port, uint16_t workers) override;
+    void Start(uint16_t port, uint16_t workers) override;
 
     // See Server.h
     void Stop() override;
